Add max_weight as PackageBuilder param
Before, the call to max_weight_for_country was delegated to the
calculator but the max_weight was computed by the package builder.
It is out of scope for the PackageBuilder since the max_weight
is determined at the calculator level. This commit remove the delegation
and make it a parameter of the process function. Also included : small
tweaks to the calculator to compute and send the parameters and changes
to the test suite.