Remove frontend controllers
The dependency on `solidus_frontend` was removed (incorrectly) in a
previous commit. If a store doesn't include it in any way, the
application will fails because the extension adds decorators for two
frontend controllers.

Those decorators are there only to rescue Spree::ShippingErrors but this
is not required because the actual errors don't bubble up to the
controllers and are rescued internally by the base calculator.

- Remove `checkout_controller_decorator`
- Remove `orders_controller_decorator`